Refactor: UploadInfoModal/index.tsx 리팩토링
장소명 검색하는 서브 모달 부분을 파일(ModalSub.tsx)로 따로 빼두었음.
